Chore: move network types into separate files
