+++ cluster_benchmark/tests/benchmark/network.rs
@@ -12,6 +12,8 @@ use openraft::error::InstallSnapshotError;
 use openraft::error::RPCError;
 use openraft::error::RaftError;
 use openraft::error::RemoteError;
+use openraft::network::RaftNetwork;
+use openraft::network::RaftNetworkFactory;
 use openraft::raft::AppendEntriesRequest;
 use openraft::raft::AppendEntriesResponse;
 use openraft::raft::InstallSnapshotRequest;
@@ -20,8 +22,6 @@ use openraft::raft::VoteRequest;
 use openraft::raft::VoteResponse;
 use openraft::Config;
 use openraft::Raft;
-use openraft::RaftNetwork;
-use openraft::RaftNetworkFactory;
 
 use crate::store::LogStore;
 use crate::store::NodeId;
