+++ examples/raft-kv-rocksdb/src/network/raft_network_impl.rs
@@ -7,6 +7,8 @@ use openraft::error::NetworkError;
 use openraft::error::RPCError;
 use openraft::error::RaftError;
 use openraft::error::RemoteError;
+use openraft::network::RaftNetwork;
+use openraft::network::RaftNetworkFactory;
 use openraft::raft::AppendEntriesRequest;
 use openraft::raft::AppendEntriesResponse;
 use openraft::raft::InstallSnapshotRequest;
@@ -14,8 +16,6 @@ use openraft::raft::InstallSnapshotResponse;
 use openraft::raft::VoteRequest;
 use openraft::raft::VoteResponse;
 use openraft::AnyError;
-use openraft::RaftNetwork;
-use openraft::RaftNetworkFactory;
 use serde::de::DeserializeOwned;
 use toy_rpc::pubsub::AckModeNone;
 use toy_rpc::Client;
