+++ openraft/src/lib.rs
@@ -78,6 +78,9 @@ mod try_as_ref;
 pub use anyerror;
 pub use anyerror::AnyError;
 pub use async_trait;
+pub use network::RPCTypes;
+pub use network::RaftNetwork;
+pub use network::RaftNetworkFactory;
 
 pub use crate::change_members::ChangeMembers;
 pub use crate::config::Config;
@@ -94,9 +97,6 @@ pub use crate::membership::EffectiveMembership;
 pub use crate::membership::Membership;
 pub use crate::membership::StoredMembership;
 pub use crate::metrics::RaftMetrics;
-pub use crate::network::RPCTypes;
-pub use crate::network::RaftNetwork;
-pub use crate::network::RaftNetworkFactory;
 pub use crate::node::BasicNode;
 pub use crate::node::EmptyNode;
 pub use crate::node::Node;
