+++ openraft/src/network/rpc_type.rs
@@ -1,0 +1,15 @@
+use std::fmt;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+pub enum RPCTypes {
+    Vote,
+    AppendEntries,
+    InstallSnapshot,
+}
+
+impl fmt::Display for RPCTypes {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
