+++ openraft/src/raft.rs
@@ -45,6 +45,7 @@ use crate::error::RaftError;
 use crate::membership::IntoNodes;
 use crate::metrics::RaftMetrics;
 use crate::metrics::Wait;
+use crate::network::RaftNetworkFactory;
 use crate::node::Node;
 use crate::storage::RaftLogStorage;
 use crate::storage::RaftStateMachine;
@@ -56,7 +57,6 @@ use crate::LogIdOptionExt;
 use crate::Membership;
 use crate::MessageSummary;
 use crate::NodeId;
-use crate::RaftNetworkFactory;
 use crate::RaftState;
 use crate::SnapshotMeta;
 use crate::StorageHelper;
