+++ tests/tests/fixtures/mod.rs
@@ -30,6 +30,8 @@ use openraft::error::RaftError;
 use openraft::error::RemoteError;
 use openraft::error::Unreachable;
 use openraft::metrics::Wait;
+use openraft::network::RaftNetwork;
+use openraft::network::RaftNetworkFactory;
 use openraft::raft::AppendEntriesRequest;
 use openraft::raft::AppendEntriesResponse;
 use openraft::raft::ClientWriteResponse;
@@ -47,8 +49,6 @@ use openraft::MessageSummary;
 use openraft::Raft;
 use openraft::RaftLogReader;
 use openraft::RaftMetrics;
-use openraft::RaftNetwork;
-use openraft::RaftNetworkFactory;
 use openraft::RaftState;
 use openraft::ServerState;
 use openraft::Vote;
