+++ tests/tests/snapshot/t33_snapshot_delete_conflict_logs.rs
@@ -4,6 +4,8 @@ use std::time::Duration;
 
 use anyhow::Result;
 use maplit::btreeset;
+use openraft::network::RaftNetwork;
+use openraft::network::RaftNetworkFactory;
 use openraft::raft::AppendEntriesRequest;
 use openraft::raft::InstallSnapshotRequest;
 use openraft::storage::RaftLogStorage;
@@ -19,8 +21,6 @@ use openraft::EntryPayload;
 use openraft::LogId;
 use openraft::Membership;
 use openraft::RaftLogReader;
-use openraft::RaftNetwork;
-use openraft::RaftNetworkFactory;
 use openraft::RaftSnapshotBuilder;
 use openraft::ServerState;
 use openraft::SnapshotPolicy;
